+++ apps/events/views.py
@@ -7,8 +7,8 @@ from rest_framework.parsers import FormParser, MultiPartParser
 from rest_framework.permissions import IsAuthenticated, IsAuthenticatedOrReadOnly
 from rest_framework.response import Response
 
+from base.permissions import IsCustomerUser, IsOrganizerUser
 from .models import Event, EventCategory, EventRegisteredUser, EventReview, Interests
-from .permissions import IsCustomerUser, IsOrganizerUser
 from .serializers import (
     EventCategorySerializer,
     EventRegisteredUserSerializer,
